Fixed issues with temp files
In the UI version, if the user created a pdf and then chose other files and created another pdf
list_images in Window class is updated to become the last set of opened images.

For example:
    User opens "1.jpg", "2.jpg" and "3.jpg"; window.list_images = ["1.jpg", "2.jpg", "3.jpg"]
    User opens "4.jpg" and "5.jpg"; window.list_images = ["4.jpg", "5.jpg"]

Because of this, in closeEvent() of window only the temporary compressed versions of the last set of files (list_images) was deleted.
This will cause extra files to stay in the temp directory after runtime.
This was fixed by having a python set for every temp file created. Window has its own temp_files set and create_pdf now returns a set that contains every temp file created in the function.
Window's temp_files and the return value from create_pdf is unioned and cleared at closeEvent.

Also, compressed files are saved as "CMP_{old_filename}" and they are compressed when the user opens them. So if the user opens images with quality x
then change it to quality y before saving, the result pdf will contain images with quality x regardless. That is because if a file exists, it is not compressed again.
To fix this, I simply changed the save name to "CMP_{quality}_{old_filename}" so every file with a different quality has unique names but if the quality is the same, it is not compressed again.